workaround(winpcap): Avoid layer 2 packet loss by filtering tcp traffic
Using a layer-2-only application, we encountered packet losses
as soon as tcp traffic was transfered on the same interface.
Checking with Wireshark it could be seen that all packets have
indeed been received by winpcap, but not handed over correctly
for some reason.
As a quick workaround this commit introduces a bpf filter with
filter string "no tcp".